Crate digits

digits

The digits crate is a linked list implementation of a score card flipper. But in this case it's with any characters you want and you can enumerate through possibilities beyond the numeric limits intrinsic in basic numerc types like u64.

Primary use case would be brute forcing character sequences.

Reexports

pub use base_custom::BaseCustom;

Structs

Digits
This struct acts similar to a full number with a custom numeric character base. But the underlying implementation is a linked list where all the methods recurse as far as need to to implement the operations.

Traits

Into