Struct digits::Digits

pub struct Digits<'a> { /* fields omitted */ }

struct Digits

This struct acts similar to a full number with a custom numeric character base. But the underlying implementation is a linked list where all the methods recurse as far as need to to implement the operations.

Methods

impl<'a> Digits<'a>

fn new<S>(mapping: &'a BaseCustom<char>, number: S) -> Digits<'a> where
S: Into<String>,

new creates a new Digits instance with the provided character set and value.

The first parameter must be a BaseCustom object which defines and maps all values. The second parameter is a string value with all valid characters from the BaseCustom set.

You can view the numbers current value at any time with to_s() or to_string().

fn to_s(&self) -> String

Gives the full value of all digits within the linked list as a String.

fn to_string(&self) -> String

Gives the full value of all digits within the linked list as a String.

fn zero(&self) -> Self

zero returns a Digits instance with value of zero and the current character mapping.

fn add(&mut self, other: Self) -> Self

Add two Digits instances together. The one the add method is called on must be mutable and modifies itself. The other is consumed.

Returns a clone of the updated Self as well.