Struct digits::Digits

pub struct Digits<'a> { /* fields omitted */ }

struct Digits

This struct acts similar to a full number with a custom numeric character base. But the underlying implementation is a linked list where all the methods recurse as far as need to to implement the operations.

Methods

impl<'a> Digits<'a>

fn new<S>(mapping: &'a BaseCustom<char>, number: S) -> Digits<'a> where
S: Into<String>,

new creates a new Digits instance with the provided character set and value.

The first parameter must be a BaseCustom object which defines and maps all values. The second parameter is a string value with all valid characters from the BaseCustom set.

You can view the numbers current value at any time with to_s() or to_string().

fn replicate(self) -> Self

replicate — alias for clone (useful for unboxing)

fn propagate<S>(&self, number: S) -> Self where
S: Into<String>,

propagate creates a new number from the same underlying numeric base

fn to_s(&self) -> String

Gives the full value of all digits within the linked list as a String.

fn to_string(&self) -> String

Gives the full value of all digits within the linked list as a String.

fn length(&self) -> usize

length returns a usize of the total linked list length.

fn zero(&self) -> Self

zero returns a Digits instance with value of zero and the current character mapping.

fn new_zero(mapping: &'a BaseCustom<char>) -> Self

new_zero returns a Digits instance with value of zero and the current character mapping.

fn is_zero(&self) -> bool

is_zero returns bool value of if the number is zero

fn pinky(&self) -> char

pinky is the smallest digit. a.k.a. current digit in the linked list. a.k.a. the right most digit. This will be a char value for that digit.

fn one(&self) -> Self

one returns a Digits instance with value of one and the current character mapping.

fn new_one(mapping: &'a BaseCustom<char>) -> Self

new_one returns a Digits instance with value of one and the current character mapping.

fn is_one(&self) -> bool

is_one returns bool value of if the number is one

fn add(&mut self, other: Self) -> Self

Add two Digits instances together. The one the add method is called on must be mutable and modifies itself. The other is consumed.

Returns a clone of the updated Self as well.

fn succ(&mut self) -> Self

succ plus one

fn pred_till_zero(&mut self) -> Self

pred_till_zero minus one unless already zero, then zero

fn pow(&mut self, pwr: Self) -> Self

pow multiplies self times the power-of given

fn mul(&mut self, other: Self) -> Self

Multiply two Digits instances together. The one the mul method is called on must be mutable and modifies itself. The other is consumed.

Returns a clone of the updated Self as well.