Struct digits::Digits

pub struct Digits<'a> { /* fields omitted */ }

This struct acts similar to a full number with a custom numeric character base. But the underlying implementation is a linked list where all the methods recurse as far as need to to implement the operations.

Methods

impl<'a> Digits<'a>

fn add(&mut self, other: Self) -> Self

Add two Digits instances together. The one the add method is called on must be mutable and modifies itself. The other is consumed.

Example

use digits::{BaseCustom,Digits};

let base10 = BaseCustom::<char>::new("0123456789".chars().collect());

let mut eleven = Digits::new(&base10, "11".to_string());
let two = Digits::new(&base10, "2".to_string());

assert_eq!(eleven.add(two).to_s(), "13");

Output

"13"

fn is_one(&self) -> bool

Returns bool value of if the number is one.

fn is_zero(&self) -> bool

Returns bool value of if the number is zero.

fn length(&self) -> usize

Returns a usize of the total linked list length.

fn mul(&mut self, other: Self) -> Self

Multiply two Digits instances together. The one the mul method is called on must be mutable and modifies itself. The other is consumed.

Example

use digits::{BaseCustom,Digits};

let base10 = BaseCustom::<char>::new("0123456789".chars().collect());

let mut eleven = Digits::new(&base10, "11".to_string());
let two = Digits::new(&base10, "2".to_string());

assert_eq!(eleven.mul(two).to_s(), "22");

Output

"22"

fn new<S>(mapping: &'a BaseCustom<char>, number: S) -> Digits<'a> where
S: Into<String>,

Creates a new Digits instance with the provided character set and value.

The first parameter must be a BaseCustom object which defines and maps all values. The second parameter is a string value with all valid characters from the BaseCustom set.

fn new_one(mapping: &'a BaseCustom<char>) -> Self

Creates a new Digits instance with value of one and the provided character mapping.

fn new_zero(mapping: &'a BaseCustom<char>) -> Self

Creates a new Digits instance with value of zero and uses the provided character mapping.

fn one(&self) -> Self

Creates a new Digits instance with value of one and uses the current character mapping.

fn pinky(&self) -> char

The “pinky” is the smallest digit a.k.a. current digit in the linked list a.k.a. the right most digit. This will be a char value for that digit.

fn pow(&mut self, pwr: Self) -> Self

Multiplies self times the power-of given Digits parameter.

Example

use digits::{BaseCustom,Digits};

let base10 = BaseCustom::<char>::new("0123456789".chars().collect());

let mut eleven = Digits::new(&base10, "11".to_string());
let two = Digits::new(&base10, "2".to_string());

assert_eq!(eleven.pow(two).to_s(), "121");

Output

"121"

fn pred_till_zero(&mut self) -> Self

Minuses one unless it's zero, then it just returns a Digits instance of zero.

fn propagate<S>(&self, number: S) -> Self where
S: Into<String>,

Creates a new Digits instance with the internal character set and given value.

The parameter is a string value with all valid characters from the BaseCustom set.

fn replicate(self) -> Self

An alias for clone. Useful for unboxing.

fn succ(&mut self) -> Self

Plus one.

fn to_s(&self) -> String

Gives the full value of all digits within the linked list as a String.

fn to_string(&self) -> String

Gives the full value of all digits within the linked list as a String.

fn zero(&self) -> Self

Creates a new Digits instance with value of zero and the current character mapping.